fix: move monitoring_info content to newly created metricsx package
